requirements.txt
# Core dependencies
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF==1.2.1
Flask-CORS==4.0.0

# Cryptography
cryptography==41.0.7
pyOpenSSL==23.3.0

# Database
SQLAlchemy>=2.0.36

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Security
bcrypt==4.1.1
pyotp==2.9.0

# API Documentation
flask-swagger-ui==4.11.1
apispec==6.3.1

# Utilities
python-dateutil==2.8.2
click==8.1.7
tabulate==0.9.0
qrcode==7.4.2

# Logging and monitoring
colorlog==6.8.0

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
